moved mergeEnv into environment.ts
